repo: Consume an entity's id on import
When importing entities, if the entity id is not already consumed
(less than or equal to its type's value in wb_id_counters) then when
the entity's id is naturally reached during entity creation, creation
will fail. This is a transient failure, as the failure will still
consume the id, but the failure can be surprising, and, if a large
block of consecutive ids are imported, frustrating to work through.

We now ensure that the entity id is consumed on import to prevent these
failures. This is done for every revision imported, before the revision
is imported. While this does impact performance more than strictly
necessary, it prevents cases where a partial failure can prevent an id
from being consumed while still causing the duplication failure on
entity create in the future. In the worst case, this will consume an id
that is not imported, and waste an id.

Bug: T289332
Change-Id: I84d1983ee88a0dd409f2f4e6f85c138095679f2e