Fix not fully dispatching Changes that added a sitelink
This was last fixed in I64ea357048c9784ec73b499ad3c8fec1fe5c8452 merged
as 1377cedb6c81c216072787ab3da8fa516c6e3710, but we broke it again when
reworking the change dispatching.

This now also adds tracking for changes with additions to the sitelinks
being dispatched, as requested by T291063. This would probably have
caught this error earlier.

This is related to a couple of open tasks (see below) and it might also
help somewhat with T297238.

Setting $this->tablesUsed in setup isn't good enough, because that way
the tables are only emptied after the tests of the class are done, not
beforehand. This is discussed in T265033.

Looking at the actual code of MediaWikiIntegrationTestCase, it seems
MediaWikiIntegrationTestCase::needsDB() is the first place in that code
that actually checks for tablesUsed. Since that method runs multiple
times, we should make sure to add our tables only once.

Bug: T233520
Bug: T298288
Bug: T291063
Bug: T295219
Change-Id: I0e189500e2a19373c9f5c9c62530627864315c7a