Avoid CAST to SIGNED for non-MySQL databases
The CAST to SIGNED introduced in
Ic3315ebd4f546462b970ddced69de9b427184005 used a syntax that is only
compatible with MySQL-like databases. Its purpose was to avoid an
unsigned integer underflow caused by a subtraction to then use in a
comparison.

The cast to a signed integer type does avoid the underflow, but the
underflow can also be avoided by performing an addition on the other
side of the comparison operator.

This does create a potential overflow issue instead, but, the overflow
issue is much less likely to occur than the underflow was, and the
overflow means catastrophic failure of the Wikibasse (from the same
overflow) is likely to occur in short order no matter what due to
storage sizes in the database.

Lastly, this change only needs to last until MediaWiki 1.37 is EOL, as
the change dispatch is replaced by a new job system in MediaWiki 1.38
and this entire class is removed.

Bug: T278510
Change-Id: Iec7dc19ca568c4ad79fb5623151e75fe944a1368