Use full database IDs when dispatching changes
PostgreSQL (by default) uses a named schema for its database. When
setting only the database name as the ID in the client wikis map,
PostgreSQL will fail to match it to the current wiki's database. This
will cause it to fail to make the jobqueuegroup with a valid domain,
and pushing the ChangeNotification jobs into the queue will fail.

Starting with MediaWiki 1.38, this change may no longer be necessary,
since the dispatchChanges maintenance script has been removed. It is
possible, however, that this change is still necessary, but in the new
queueing code.

Change-Id: I167830f5979d01970b0991124046e1a9f2756d69